Move the Parser effect into its own module.
