+++ src/Path/Parser/REPL.hs
@@ -2,6 +2,7 @@
 module Path.Parser.REPL where
 
 import Control.Applicative (Alternative(..))
+import Control.Effect.Parser
 import Control.Effect.Reader
 import Path.Parser
 import qualified Path.Parser.Module as M
